% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_consistency_check.R
\name{check_that_species_are_tested_for_more_than_one_concentration}
\alias{check_that_species_are_tested_for_more_than_one_concentration}
\title{Check that species are tested for more than one concentration}
\usage{
check_that_species_are_tested_for_more_than_one_concentration(wide_dataset)
}
\arguments{
\item{dataset}{The loaded dataset. Must contain only one contaminant.}

\item{concentration_col_name}{The name of the column containing the concentrations.}

\item{time_col_name}{The name of the column containing the time stamps.}

\item{n_alive_col_name}{The name of the column containing the number of animals alive.}

\item{species_identifier_col_name}{The name of the column containing a single identifier per species.}

\item{test_id_col_name}{The name of the column containing a identifier to dis-ambiguate several tests for the same species.}

\item{keep_only}{Choose whether to keep the whole dataset or only the necessary columns}
}
\value{
The dataset with columns \code{number_alive}, \code{concentration}, \code{time} and \code{species} and eventually the \code{test_id}.
}
\description{
Check that species are tested for more than one concentration
}
\examples{
library(tidyverse)
dataset = read_csv('../data/RTT_dataV2.csv') \%>\%
    setNames(nm = sapply(X = names(.), FUN = function(xx) gsub(' ','_',xx)))

Finish when some more pre-processing is done

dataset \%>\%
   rename()
}
