% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_consistency_check.R
\name{check_that_the_number_of_survivors_does_not_increase}
\alias{check_that_the_number_of_survivors_does_not_increase}
\title{Check that the number of survivors does not increase}
\usage{
check_that_the_number_of_survivors_does_not_increase(raw_wide_dataset,
  time_col_names, time_stamps)
}
\arguments{
\item{raw_wide_dataset}{The loaded dataset.}

\item{time_col_names}{The name of the column containing number of animals alive at for a specific time.}
}
\value{
\code{TRUE} if the number of survivors does not increase, else \code{FALSE}
}
\description{
Check that the number of survivors does not increase
}
