% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_closest_strings.R
\name{find_closest_strings}
\alias{find_closest_strings}
\title{Find the closest pairs of strings from a vector of strings.}
\usage{
find_closest_strings(strings_, quantile_ = 0.01)
}
\arguments{
\item{strings_}{A vector of strings. If the vector contain duplicates strings, they will be removed.}

\item{quantile_}{Proportion of the closest pairs of strings to return. For quantile_ = 0.1, the best 10 percent pairs are returned.}
}
\value{
A matrix containing the "quantile_" closest pairs of strings.
}
\description{
This is using the Jaro-Winckler distance to identify the closest strings,
see https://journal.r-project.org/archive/2014-1/loo.pdf
}
\examples{
library(tidyverse)
read_csv('../data/RTT_dataV2.csv') \%>\%
    .$Identifier \%>\%
    find_closest_strings

}
