% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_functions.R
\name{fit_prepared_dataset_jags}
\alias{fit_prepared_dataset_jags}
\title{# library(rstan)
rstan::rstan_options(auto_write = TRUE)
options(mc.cores = parallel::detectCores())}
\usage{
fit_prepared_dataset_jags(prepared_dataset)
}
\description{

}
\details{

}
