% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_consistency_check.R
\name{get_mask_for_unsorted_rows}
\alias{get_mask_for_unsorted_rows}
\title{Get boolean mask for unsorted rows}
\usage{
get_mask_for_unsorted_rows(raw_wide_dataset, time_col_names, time_stamps)
}
\arguments{
\item{raw_wide_dataset}{The loaded dataset.}

\item{time_col_names}{The name of the column containing number of animals alive at for a specific time.}

\item{time_stamps}{The dates at which the number of animals alive is measured, in the same order as time_col_names}

\item{concentration_col_name}{The name of the column containing the concentrations.}
}
\value{
A mask vector with \code{TRUE} for unsorted rows and \code{false} for sorted rows.
}
\description{
Get boolean mask for unsorted rows
}
\examples{

}
