% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_consistency_check.R
\name{get_rows_where_the_number_of_survivors_increases}
\alias{get_rows_where_the_number_of_survivors_increases}
\title{Get the rows where the number of survivors increases}
\usage{
get_rows_where_the_number_of_survivors_increases(
  raw_wide_dataset,
  time_columns,
  time_stamps
)
}
\arguments{
\item{raw_wide_dataset}{The loaded dataset.}

\item{time_stamps}{The dates at which the number of animals alive is measured, in the same order as time_col_names}
}
\value{
The rows where the number of survivors increases.
}
\description{
Get the rows where the number of survivors increases
}
