% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highest_posterior_density_interval.R
\name{HDIofGrid}
\alias{HDIofGrid}
\title{Kruschke, J. (2014). Doing Bayesian data analysis: A tutorial introduction with R, 2nd Ed. Academic Press.}
\usage{
HDIofGrid(probMassVec, credMass = 0.95)
}
\arguments{
\item{probMassVec}{is a vector of probability masses at each grid point.}

\item{credMass}{is the desired mass of the HDI region.}
}
\value{
value: A list with components:
indices is a vector of indices that are in the HDI
mass is the total mass of the included indices
height is the smallest component probability mass in the HDI
}
\description{
Kruschke, J. (2014). Doing Bayesian data analysis: A tutorial introduction with R, 2nd Ed. Academic Press.
}
