% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_dataset.R
\name{rename_dataset}
\alias{rename_dataset}
\title{Specify which are the interesting column of a dataset and rename them for standardized use.}
\usage{
rename_dataset(dataset, concentration_col_name, time_col_name, n_alive_col_name,
  species_identifier_col_name, test_id_col_name = NULL, keep_only = FALSE)
}
\arguments{
\item{dataset}{The loaded dataset. Must contain only one contaminant.}

\item{concentration_col_name}{The name of the column containing the concentrations.}

\item{time_col_name}{The name of the column containing the time stamps.}

\item{n_alive_col_name}{The name of the column containing the number of animals alive.}

\item{species_identifier_col_name}{The name of the column containing a single identifier per species.}

\item{test_id_col_name}{The name of the column containing a identifier to dis-ambiguate several tests for the same species.}

\item{keep_only}{Choose whether to keep the whole dataset or only the necessary columns}
}
\value{
The dataset with columns \code{number_alive}, \code{concentration}, \code{time} and \code{species} and eventually the \code{test_id}.
}
\description{
The argument \code{keep_only} is intended to make the function more versatile.
}
\examples{
library(tidyverse)
dataset = read_csv('../data/RTT_dataV2.csv') \%>\%
    setNames(nm = sapply(X = names(.), FUN = function(xx) gsub(' ','_',xx)))

Finish when some more pre-processing is done

dataset \%>\%
   rename()
}
